\name{OpasnetBaseUtils-package}
\alias{OpasnetBaseUtils-package}
\alias{OpasnetBaseUtils}
\docType{package}
\title{Package for dealing with Opasnet Base data}
\description{
  Opasnet (\url{http://www.opasnet.org}) database tools.
}
\details{
\tabular{ll}{
Package: \tab OpasnetBaseUtils\cr
Type: \tab Package\cr
Version: \tab 0.8.5\cr
Date: \tab 2011-09-13\cr
License: \tab CC BY-SA\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Teemu R\cr
Einari H, \email{einari.happonen@thl.fi}
}
\keyword{ package }